Cleaned up code, added logging, and added error handling
